refactor(web): FilterPanel を別ファイルへ分離し what コメントを削除
page.tsx をページコンポーネントのみに保つため FilterPanel を filter-panel.tsx へ抽出。
あわせて route.tsx の what コメントを削除した。
